fix: extract updateSchedule to admin-helpers to stay under 300-line limit
